mapvote 1.0.1
Fixed undo key for controllers.
Controllers are now fully supported

Removed the frag/+frag button for the undo option as it wasn't working for controllers

Moved the use/+activate key from the select option (by default F on PC and Square on Playstation controllers) to the undo option instead.

Updated the README to add controller support information and screenshot