mapvote 2.0.0
Added support for zombies mode.
Using the mapvote_maps dvar players can set a display name, the map name and the cfg name

Added mapvote_limits_maps dvar
Added mapvote_limits_modes dvar
Added mapvote_limits_max dvar
Users can now define maps and/or modes limits to have a precise amount of maps and/or modes on screen even if the available amount is above the max limit. The max limit is now 12 but can be lowered using a dvar

Changed right click listener wait time from 0.35s to 0.25s for a better user experience

Hide the map or mode when there's only one in the mapvote.
It will still rotate to it but it's hidden so that the menu looks cleaner. The menu is centered accordingly

Fixed the menu centering calculation, it's now really centered

Removed the need to write map names in a certain way to work, users can now write map names the way they want