chat_commands 1.0.0
This is a port of my MW3 chat_commands script (version 1.3.2)
https://github.com/Resxt/Plutonium-IW5-Scripts/commit/ba1322b933e5afcbf9b7dd1348d26485ceae3afa

The array declarations such as array = ["value"] were changed to use the array function

The tell function (player chat) was changed IPrintLnBold. More explanation in the README

Support for clantags was added. Clantags were changing the player's name which would then make the script fail to find the player when typing his name. This has been tested and fixed. More explanation in the README

Errors and util functions related to camo were removed. The way camo works seems to be really different than MW3 so work is required to implement these again