chat_commands 1.0.4
Only create variables for the port currently in-use
For example: there's no need to create level.chat_commands["4977"] if the currently used port is 4976. This simply adds a bunch of variables and child variables for no reason, this was a flawed implementation that create an array per port on every server. Now it will only create an array for its own port as expected

Removed 3074 from the default ports
This port is not actively used by net_port, this was just a misunderstanding on my part